Adding support for build.rs scripts
Fixes: #51
